app/requirements.txt
Django~=5.0.4
tensorflow~=2.16.1
opencv-python-headless
pillow